+++ Proxy/proxy_threads/pt_http_utl.h
@@ -18,5 +18,5 @@ void pt_http_write_destroy();
 
 int pt_http_read(char** in_buf); // Returns 0 if timeout or actual buf len (LONG GET), < 0 if error (-errno)
 
-int pt_http_write(char* buf, unsigned int len, char** resp, unsigned int* resp_len); //Returns 0 if timeout or error, 1 if OK
+int pt_http_write(char* buf, size_t len, char** resp, size_t* resp_len); //Returns 0 if timeout or error, 1 if OK
 #endif //PT_HTTP_UTL_H
